refactor(#1491): split the coverage census out of the lane file (campsite rule)
The lane file crossed the 1500-line test target (1508), so the coverage-census
half moves to `support/issue_1491_coverage_census.rs` (298 lines) — the
`Excluded`/`NOT_COMPARABLE` classification, `committed_fixture_coverage_census`
and `every_declarable_shape_is_one_the_golden_reader_refuses`. A different
question from egress parity: which fixtures the lane is accountable for.

Declared as a CHILD module of the lane, so it reads the same `CASES` and
`repo_root` the comparison uses and there is no second copy of the case table to
drift. Lane file 1233 lines; every file in the lane is under target. No behaviour
change: 126 passed / 0 failed, same tests.
